+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## [0.1.4]
+
+* Added "isPausingSendNotifications" to INotifyPropertyChanged and NotifyPropertyChanged.
+* Added "isBlank()" and changed impl. of "isEmpty()" within string_utils.dart
+
+## [0.1.3]
+
+* fixes AppNavigator.
+
 ## [0.1.2]
 
 * Readme corrections.